+++ starlette_context/middleware/context_middleware.py
@@ -9,14 +9,15 @@ from starlette.requests import Request
 from starlette.responses import Response
 
 from starlette_context import _request_scope_context_storage
+from starlette_context.middleware.mixin import StarletteContextMiddlewareMixin
 from starlette_context.plugins import Plugin
 from starlette_context.errors import (
     ConfigurationError,
-    MiddleWareValidationError,
+    StarletteContextException,
 )
 
 
-class ContextMiddleware(BaseHTTPMiddleware):
+class ContextMiddleware(BaseHTTPMiddleware, StarletteContextMiddlewareMixin):
     """
     Middleware that creates empty context for request it's used on. If not
     used, you won't be able to use context object.
@@ -28,18 +29,18 @@ class ContextMiddleware(BaseHTTPMiddleware):
     def __init__(
         self,
         plugins: Optional[Sequence[Plugin]] = None,
-        default_error_response: Response = Response(status_code=400),
-        *args,
         **kwargs,
     ) -> None:
-        super().__init__(*args, **kwargs)
+        BaseHTTPMiddleware.__init__(
+            self, app=kwargs["app"], dispatch=kwargs.get("dispatch")
+        )
+        StarletteContextMiddlewareMixin.__init__(self, **kwargs)
         for plugin in plugins or ():
             if not isinstance(plugin, Plugin):
                 raise ConfigurationError(
                     f"Plugin {plugin} is not a valid instance"
                 )
         self.plugins = plugins or ()
-        self.error_response = default_error_response
 
     async def set_context(self, request: Request) -> dict:
         """
@@ -59,19 +60,15 @@ class ContextMiddleware(BaseHTTPMiddleware):
         try:
             context = await self.set_context(request)
             token: Token = _request_scope_context_storage.set(context)
-        except MiddleWareValidationError as e:
-            if e.error_response:
-                error_response = e.error_response
-            else:
-                error_response = self.error_response
-            return error_response
 
-        try:
-            response = await call_next(request)
-            for plugin in self.plugins:
-                await plugin.enrich_response(response)
+            try:
+                response = await call_next(request)
+                for plugin in self.plugins:
+                    await plugin.enrich_response(response)
+            finally:
+                _request_scope_context_storage.reset(token)
 
-        finally:
-            _request_scope_context_storage.reset(token)
+        except StarletteContextException as e:
+            response = await self.create_response_from_exception(request, e)
 
         return response