+++ starlette_context/middleware/raw_middleware.py
@@ -2,24 +2,25 @@ from contextvars import Token
 from typing import Optional, Sequence, Union
 
 from starlette.requests import HTTPConnection, Request
-from starlette.responses import Response
 from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
 from starlette_context import _request_scope_context_storage
+from starlette_context.middleware.mixin import StarletteContextMiddlewareMixin
 from starlette_context.plugins import Plugin
 from starlette_context.errors import (
     ConfigurationError,
-    MiddleWareValidationError,
+    StarletteContextException,
 )
 
 
-class RawContextMiddleware:
+class RawContextMiddleware(StarletteContextMiddlewareMixin):
     def __init__(
         self,
         app: ASGIApp,
         plugins: Optional[Sequence[Plugin]] = None,
-        default_error_response: Response = Response(status_code=400),
+        **kwargs,
     ) -> None:
+        super().__init__(**kwargs)
         self.app = app
         for plugin in plugins or ():
             if not isinstance(plugin, Plugin):
@@ -28,7 +29,6 @@ class RawContextMiddleware:
                 )
 
         self.plugins = plugins or ()
-        self.error_response = default_error_response
 
     async def set_context(
         self, request: Union[Request, HTTPConnection]
@@ -70,29 +70,30 @@ class RawContextMiddleware:
 
         try:
             context = await self.set_context(request)
-        except MiddleWareValidationError as e:
-            if e.error_response is not None:
-                error_response = e.error_response
-            else:
-                error_response = self.error_response
+            token: Token = _request_scope_context_storage.set(context)
 
-            message_head = {
+            try:
+                await self.app(scope, receive, send_wrapper)
+            finally:
+                _request_scope_context_storage.reset(token)
+
+        except StarletteContextException as e:
+            response = await self.create_response_from_exception(request, e)
+
+            message_head: Message = {
                 "type": "http.response.start",
-                "status": error_response.status_code,
-                "headers": error_response.raw_headers,
+                "status": response.status_code,
+                "headers": [
+                    (k.encode(), v.encode())
+                    for k, v in response.headers.items()
+                ],
             }
+
             await send(message_head)
 
-            message_body = {
+            message_body: Message = {
                 "type": "http.response.body",
-                "body": error_response.body,
+                "body": response.body,
             }
             await send(message_body)
             return
-
-        token: Token = _request_scope_context_storage.set(context)
-
-        try:
-            await self.app(scope, receive, send_wrapper)
-        finally:
-            _request_scope_context_storage.reset(token)